Renaming one file.
